* io.c (rb_write_error2): fwrite() returns ssize_t.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@33884 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ io.c
@@ -6599,7 +6599,7 @@ void
 rb_write_error2(const char *mesg, long len)
 {
     if (rb_stderr == orig_stderr || RFILE(orig_stderr)->fptr->fd < 0) {
-	int unused;
+	ssize_t unused;
 	unused = fwrite(mesg, sizeof(char), len, stderr);
     }
     else {
